Add emojis that work with colemak layout
The problem here is that I'm setting Colemak on my OS not on the
keyboard directly. I'm doing this because I want to have colemak working
on other keyboards as well (my kinesis and also the laptop keyboard).

The hex was being sent and the OS was changing it to the colemak
respective letters. In this case, the "f" of the emoji Hex would become a
"t" and it would not work.

What I'm doing here is to change it to a different "hex" and let the OS
change and get the right emoji. It's a workaround and it doesn't work
for all emojis but it's OK for now.